bug(app): relocated app.js into app dir [start #165407939]
